refactor: change assets
