+++ assets/atlas/enemy/knight/knight.atlas
@@ -1,0 +1,137 @@
+knight.png
+size:2048,256
+repeat:none
+spr_attack
+index:6
+bounds:692,122,111,132
+spr_attack
+index:4
+bounds:805,71,81,183
+spr_attack
+index:5
+bounds:888,71,66,183
+spr_attack
+index:7
+bounds:956,125,84,129
+spr_attack
+index:9
+bounds:1042,122,66,132
+spr_attack
+index:8
+bounds:1110,125,66,129
+spr_attack
+index:3
+bounds:1178,155,87,99
+spr_attack
+index:1
+bounds:1611,155,84,99
+spr_attack
+index:0
+bounds:956,27,84,96
+spr_attack
+index:12
+bounds:1042,27,69,93
+spr_attack
+index:2
+bounds:1199,54,84,99
+spr_attack
+index:10
+bounds:1371,39,69,114
+spr_attack
+index:11
+bounds:1442,63,69,90
+spr_die
+index:8
+bounds:2,106,256,148
+spr_die
+index:7
+bounds:260,106,240,148
+spr_die
+index:6
+bounds:502,126,188,128
+spr_die
+index:5
+bounds:2,16,120,88
+spr_die
+index:1
+bounds:1513,41,44,112
+spr_die
+index:2
+bounds:1559,81,48,72
+spr_die
+index:4
+bounds:1609,97,64,56
+spr_die
+index:3
+bounds:1985,113,48,40
+spr_idle
+bounds:296,5,84,99
+spr_run
+index:0
+bounds:296,5,84,99
+spr_run
+index:5
+bounds:1267,155,84,99
+spr_run
+index:4
+bounds:1267,155,84,99
+spr_run
+index:6
+bounds:1267,155,84,99
+spr_run
+index:11
+bounds:1353,155,84,99
+spr_run
+index:16
+bounds:1439,155,84,99
+spr_run
+index:2
+bounds:1525,155,84,99
+spr_run
+index:7
+bounds:1697,155,84,99
+spr_run
+index:13
+bounds:1783,155,84,99
+spr_run
+index:10
+bounds:1869,155,84,99
+spr_run
+index:1
+bounds:1955,155,84,99
+spr_run
+index:9
+bounds:124,8,84,96
+spr_run
+index:15
+bounds:210,5,84,99
+spr_run
+index:14
+bounds:210,5,84,99
+spr_run
+index:12
+bounds:382,5,84,99
+spr_run
+index:3
+bounds:1113,24,84,99
+spr_run
+index:8
+bounds:1285,57,84,96
+spr_spawn
+index:5
+bounds:1442,2,60,59
+spr_spawn
+index:2
+bounds:1675,94,60,59
+spr_spawn
+index:4
+bounds:1737,94,60,59
+spr_spawn
+index:1
+bounds:1799,94,60,59
+spr_spawn
+index:3
+bounds:1861,94,60,59
+spr_spawn
+index:0
+bounds:1923,94,60,59
