--- assets/atlas/enemy/mushroom/mushroom.atlas
@@ -1,141 +1,0 @@
-mushroom.png
-size:2048,256
-repeat:none
-spr_attack
-index:4
-bounds:450,126,80,64
-spr_attack
-index:9
-bounds:532,60,80,64
-spr_attack
-index:1
-bounds:614,60,80,64
-spr_attack
-index:6
-bounds:778,60,80,64
-spr_attack
-index:3
-bounds:942,60,80,64
-spr_attack
-index:8
-bounds:1024,60,80,64
-spr_attack
-index:5
-bounds:1188,60,80,64
-spr_attack
-index:2
-bounds:1270,60,80,64
-spr_attack
-index:7
-bounds:1352,60,80,64
-spr_die
-index:3
-bounds:2,97,110,93
-spr_die
-index:8
-bounds:2,2,110,93
-spr_die
-index:0
-bounds:2,2,110,93
-spr_die
-index:10
-bounds:2,2,110,93
-spr_die
-index:9
-bounds:2,2,110,93
-spr_die
-index:5
-bounds:114,97,110,93
-spr_die
-index:2
-bounds:114,2,110,93
-spr_die
-index:7
-bounds:226,97,110,93
-spr_die
-index:4
-bounds:226,2,110,93
-spr_die
-index:1
-bounds:338,97,110,93
-spr_die
-index:6
-bounds:338,2,110,93
-spr_hit
-index:4
-bounds:614,126,80,64
-spr_attack
-index:0
-bounds:614,126,80,64
-spr_hit
-index:0
-bounds:614,126,80,64
-spr_idle
-index:6
-bounds:614,126,80,64
-spr_spawn
-index:0
-bounds:614,126,80,64
-spr_spawn
-index:4
-bounds:614,126,80,64
-spr_hit
-index:1
-bounds:860,126,80,64
-spr_spawn
-index:1
-bounds:860,126,80,64
-spr_hit
-index:3
-bounds:1106,126,80,64
-spr_spawn
-index:3
-bounds:1106,126,80,64
-spr_idle
-index:5
-bounds:450,60,80,64
-spr_idle
-index:2
-bounds:696,126,80,64
-spr_idle
-index:4
-bounds:942,126,80,64
-spr_idle
-index:3
-bounds:942,126,80,64
-spr_idle
-index:1
-bounds:1106,60,80,64
-spr_idle
-index:0
-bounds:1434,126,80,64
-spr_run
-index:5
-bounds:532,126,80,64
-spr_run
-index:2
-bounds:778,126,80,64
-spr_run
-index:7
-bounds:860,60,80,64
-spr_run
-index:4
-bounds:1024,126,80,64
-spr_run
-index:1
-bounds:1188,126,80,64
-spr_run
-index:6
-bounds:1270,126,80,64
-spr_run
-index:3
-bounds:1352,126,80,64
-spr_run
-index:0
-bounds:1434,60,80,64
-spr_spawn
-index:2
-bounds:696,60,80,64
-spr_hit
-index:2
-bounds:696,60,80,64
